big-project/requirements.txt
absl-py
yaml
